pyproject.toml
[tool.poetry]
name = "operator-by-phone"
version = "0.1.0"
description = "Сервис определения оператора и региона по номеру телефона"
authors = ["Artasov <ivanhvalevskey@gmail.com>"]
license = "MIT"
readme = "README.md"
homepage = "https://localhost:8000"
repository = "https://github.com/Artasov/operator-by-phone"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
Django = "^5.1.8"
adjango = "^0.5.5"
djangorestframework = "3.16.0"
celery = "^5.5.1"
redis = "^5.2.1"
gunicorn = "^23.0.0"
whitenoise = "^6.9.0"
django-celery-beat = "^2.7.0"
requests = "^2.32.3"
python-dotenv = "^1.1.0"
psycopg2-binary = "^2.9.10"

[tool.poetry.group.dev.dependencies]
pytest = "^7.4.2"
pytest-django = "^4.11.1"
flake8 = "^5.0.4"
black = "^23.12.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
